Update chartfooter.php
'break' on -1 seems to fix an issue where the markings were not displayed for the most recent 'boiler on' time - there may be a better way of doing this, but it seems to work. 
The "if (is_null)..." allows the most current boiler on time to be displayed, even if it's still running (ie the stop time is NULL) by 'faking' the stop time as 'now'.  
Switching $boiler_start & $boiler_stop in the from/to doesn't appear to make any difference to anything - it displays fine either way!?!  Flot is weird, sometimes.